Cloudformation job exits with error if instance exists on vpc
[#101717890](https://www.pivotaltracker.com/story/show/101717890)

+++ ci/tasks/cloudformation.sh
@@ -14,8 +14,17 @@ export AWS_DEFAULT_REGION=${region_name}
 
 stack_name="aws-cpi-stack"
 
+stack_info=$(get_stack_info $stack_name)
+vpc_id=$(get_stack_info_of "$stack_info" "vpcid")
+instances=$(aws ec2 describe-instances --query "Reservations[*].Instances[*].InstanceId[]" --filters "Name=vpc-id,Values=${vpc_id}")
+
+if [ ! -z "$instances" ] ; then
+  echo "Error: Alive instances found on ${vpc_id}: ${instances}"
+  exit 1
+fi
+
 cmd="aws cloudformation delete-stack --stack-name ${stack_name}"
-echo "Running: #{cmd}"; ${cmd}
+echo "Running: ${cmd}"; ${cmd}
 
 while true; do
   stack_status=$(get_stack_status $stack_name)
@@ -32,12 +41,11 @@ while true; do
   fi
 done
 
-cmd<<EOF
-aws cloudformation create-stack \
+cmd="aws cloudformation create-stack \
     --stack-name      ${stack_name} \
-    --template-body   file:///${PWD}/bosh-cpi-release/ci/assets/cloudformation.templateEOF
-EOF
-echo "Running: #{cmd}"; ${cmd}
+    --template-body   file:///${PWD}/bosh-cpi-release/ci/assets/cloudformation.template"
+
+echo "Running: ${cmd}"; ${cmd}
 while true; do
   stack_status=$(get_stack_status $stack_name)
   echo "StackStatus ${stack_status}"
